# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/charging-point":{"post":{"operationId":"ChargePointController_create","parameters":[],"responses":{"201":{"description":""}},"tags":["ChargePoint"]},"get":{"operationId":"ChargePointController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["ChargePoint"]}},"/charging-point/{id}":{"get":{"operationId":"ChargePointController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ChargePoint"]},"patch":{"operationId":"ChargePointController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateChargePointDto"}}}},"responses":{"200":{"description":""}},"tags":["ChargePoint"]},"delete":{"operationId":"ChargePointController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ChargePoint"]}},"/stations":{"post":{"operationId":"StationController_createStation","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateStationDto"}}}},"responses":{"201":{"description":"The station has been successfully created."},"400":{"description":"Invalid input."},"409":{"description":"Station already exists."}},"security":[{"bearer":[]}],"summary":"Create a new station","tags":["stations"]},"get":{"operationId":"StationController_findAllStations","parameters":[],"responses":{"200":{"description":"Returns all stations matching the query parameters."}},"security":[{"bearer":[]}],"summary":"Get all stations","tags":["stations"]}},"/stations/{id}":{"get":{"operationId":"StationController_findStationById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns the station with the specified id."},"404":{"description":"Station not found."}},"security":[{"bearer":[]}],"summary":"Get a station by id","tags":["stations"]},"put":{"operationId":"StationController_updateStation","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStationDto"}}}},"responses":{"200":{"description":"The station has been successfully updated."},"400":{"description":"Invalid input."},"404":{"description":"Station not found."}},"security":[{"bearer":[]}],"summary":"Update a station","tags":["stations"]},"delete":{"operationId":"StationController_deleteStation","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The station has been successfully deleted."},"404":{"description":"Station not found."}},"security":[{"bearer":[]}],"summary":"Delete a station","tags":["stations"]}},"/locations":{"post":{"operationId":"LocationController_createLocation","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLocationDto"}}}},"responses":{"201":{"description":"The location has been successfully created."}},"security":[{"bearer":[]}],"summary":"Create a new location","tags":["locations"]},"get":{"operationId":"LocationController_findAllLocations","parameters":[],"responses":{"200":{"description":"Returns all locations matching the query parameters."}},"security":[{"bearer":[]}],"summary":"Get all locations","tags":["locations"]}},"/locations/{id}":{"get":{"operationId":"LocationController_findLocationById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns the location with the specified id."}},"security":[{"bearer":[]}],"summary":"Get a location by id","tags":["locations"]},"put":{"operationId":"LocationController_updateLocation","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLocationDto"}}}},"responses":{"200":{"description":"The location has been successfully updated."}},"security":[{"bearer":[]}],"summary":"Update a location","tags":["locations"]},"delete":{"operationId":"LocationController_deleteLocation","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The location has been successfully deleted."}},"security":[{"bearer":[]}],"summary":"Delete a location","tags":["locations"]}},"/users":{"post":{"operationId":"UserController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"409":{"description":"Email already exists"}},"security":[{"bearer":[]}],"summary":"Create a new user","tags":["users"]},"get":{"operationId":"UserController_findAllUsers","parameters":[{"name":"skip","required":true,"in":"query","schema":{"type":"number"}},{"name":"take","required":true,"in":"query","schema":{"type":"number"}},{"name":"search","required":true,"in":"query","schema":{"type":"string"}},{"name":"role","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}},"security":[{"bearer":[]}],"summary":"Get all users","tags":["users"]}},"/users/{id}":{"get":{"operationId":"UserController_findUserById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"404":{"description":"User not found"}},"security":[{"bearer":[]}],"summary":"Get a user by id","tags":["users"]},"put":{"operationId":"UserController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"404":{"description":"User not found"}},"security":[{"bearer":[]}],"summary":"Update a user","tags":["users"]},"delete":{"operationId":"UserController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"404":{"description":"User not found"}},"security":[{"bearer":[]}],"summary":"Delete a user","tags":["users"]}},"/users/{id}/change-password":{"put":{"operationId":"UserController_changePassword","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"404":{"description":"User not found"},"409":{"description":"Current password is incorrect"}},"security":[{"bearer":[]}],"summary":"Change user password","tags":["users"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User successfully registered"},"409":{"description":"Email already exists"}},"summary":"Register a new user","tags":["auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Login successful"},"401":{"description":"Invalid credentials"}},"summary":"Login with email and password","tags":["auth"]}},"/auth/refresh":{"post":{"operationId":"AuthController_refreshToken","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"Token refreshed successfully"},"401":{"description":"Invalid refresh token"}},"summary":"Refresh access token","tags":["auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":"Logout successful"}},"summary":"Logout user","tags":["auth"]}},"/auth/request-password-reset":{"post":{"operationId":"AuthController_requestPasswordReset","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestPasswordResetDto"}}}},"responses":{"200":{"description":"Reset email sent if user exists"}},"summary":"Request password reset email","tags":["auth"]}},"/auth/me":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":"User profile retrieved"}},"summary":"Get current user profile","tags":["auth"]}},"/auth/google":{"get":{"operationId":"AuthController_googleAuth","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/auth/google/callback":{"get":{"operationId":"AuthController_googleAuthCallback","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/auth/github":{"get":{"operationId":"AuthController_githubAuth","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/auth/github/callback":{"get":{"operationId":"AuthController_githubAuthCallback","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/mail/test/project-invitation":{"post":{"operationId":"MailController_sendTestProjectInvitation","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestProjectInvitationDto"}}}},"responses":{"200":{"description":"Test email sent successfully"},"400":{"description":"Invalid email data provided"}},"security":[{"bearer":[]}],"summary":"Send a test project invitation email","tags":["mail"]}},"/mail/test/password-reset":{"post":{"operationId":"MailController_sendTestPasswordReset","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestPasswordResetDto"}}}},"responses":{"200":{"description":"Test email sent successfully"}},"security":[{"bearer":[]}],"summary":"Send a test password reset email","tags":["mail"]}},"/mail/verify-connection":{"get":{"operationId":"MailController_verifyConnection","parameters":[],"responses":{"200":{"description":"Connection status"}},"security":[{"bearer":[]}],"summary":"Verify SMTP connection","tags":["mail"]}},"/mail/preview/project-invitation":{"get":{"operationId":"MailController_previewProjectInvitation","parameters":[{"name":"to","required":true,"in":"query","schema":{"example":"skbhati199@gmail.com","type":"string"}},{"name":"inviterName","required":true,"in":"query","schema":{"example":"sk","type":"string"}},{"name":"projectName","required":true,"in":"query","schema":{"example":"Project Demo","type":"string"}},{"name":"role","required":true,"in":"query","schema":{"example":"PROJECT_ADMIN","type":"string"}},{"name":"expiresAt","required":true,"in":"query","schema":{"format":"date-time","example":"2023-01-01","type":"string"}}],"responses":{"200":{"description":"Returns the rendered HTML template"}},"security":[{"bearer":[]}],"summary":"Preview project invitation email template","tags":["mail"]}},"/mail/preview/password-reset":{"get":{"operationId":"MailController_previewPasswordReset","parameters":[],"responses":{"200":{"description":"Returns the rendered HTML template"}},"security":[{"bearer":[]}],"summary":"Preview password reset email template","tags":["mail"]}},"/mail/verify":{"post":{"operationId":"MailController_verifyEmail","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["mail"]}},"/mail/resend":{"post":{"operationId":"MailController_resendOTP","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["mail"]}},"/organizations":{"post":{"operationId":"OrgController_createOrganization","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrgDto"}}}},"responses":{"201":{"description":"Organization successfully created"},"409":{"description":"Organization with this email already exists"}},"security":[{"bearer":[]}],"summary":"Create a new organization","tags":["organizations"]},"get":{"operationId":"OrgController_findAllOrganizations","parameters":[],"responses":{"200":{"description":"Returns all organizations"}},"security":[{"bearer":[]}],"summary":"Get all organizations","tags":["organizations"]}},"/organizations/{id}":{"get":{"operationId":"OrgController_findOrganizationById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"includeRelations","required":true,"in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":"Returns the organization"},"404":{"description":"Organization not found"}},"security":[{"bearer":[]}],"summary":"Get organization by id","tags":["organizations"]},"put":{"operationId":"OrgController_updateOrganization","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrgDto"}}}},"responses":{"200":{"description":"Organization updated"},"404":{"description":"Organization not found"}},"security":[{"bearer":[]}],"summary":"Update an organization","tags":["organizations"]},"delete":{"operationId":"OrgController_deleteOrganization","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Organization deleted"},"404":{"description":"Organization not found"},"409":{"description":"Cannot delete organization with dependencies"}},"security":[{"bearer":[]}],"summary":"Delete an organization","tags":["organizations"]}},"/organizations/{id}/users/{userId}":{"post":{"operationId":"OrgController_addUserToOrganization","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"User added to organization"},"404":{"description":"Organization or user not found"}},"security":[{"bearer":[]}],"summary":"Add user to organization","tags":["organizations"]},"delete":{"operationId":"OrgController_removeUserFromOrganization","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"User removed from organization"},"404":{"description":"Organization or user not found"}},"security":[{"bearer":[]}],"summary":"Remove user from organization","tags":["organizations"]}},"/projects":{"post":{"operationId":"ProjectController_createProject","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"201":{"description":"Project successfully created"}},"security":[{"bearer":[]}],"summary":"Create a new project","tags":["projects"]},"get":{"operationId":"ProjectController_findAllProjects","parameters":[],"responses":{"200":{"description":"Returns all accessible projects"}},"security":[{"bearer":[]}],"summary":"Get all projects","tags":["projects"]}},"/projects/{id}":{"get":{"operationId":"ProjectController_findProjectById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns the project"},"404":{"description":"Project not found"}},"security":[{"bearer":[]}],"summary":"Get project by id","tags":["projects"]},"put":{"operationId":"ProjectController_updateProject","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectDto"}}}},"responses":{"200":{"description":"Project updated"},"404":{"description":"Project not found"}},"security":[{"bearer":[]}],"summary":"Update a project","tags":["projects"]}},"/projects/{id}/members":{"post":{"operationId":"ProjectController_addProjectMember","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddProjectMemberDto"}}}},"responses":{"200":{"description":"Member added to project"},"404":{"description":"Project not found"}},"security":[{"bearer":[]}],"summary":"Add member to project","tags":["projects"]}},"/projects/{id}/members/{memberId}":{"delete":{"operationId":"ProjectController_removeProjectMember","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"memberId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Member removed from project"},"404":{"description":"Project member not found"}},"security":[{"bearer":[]}],"summary":"Remove member from project","tags":["projects"]}},"/projects/{id}/members/{memberId}/role":{"put":{"operationId":"ProjectController_updateMemberRole","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"memberId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Member role updated"},"404":{"description":"Project member not found"}},"security":[{"bearer":[]}],"summary":"Update member role in project","tags":["projects"]}},"/project-invitations":{"post":{"operationId":"ProjectInvitationController_createInvitation","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectInvitationDto"}}}},"responses":{"201":{"description":"Invitation successfully created and email sent"}},"security":[{"bearer":[]}],"summary":"Create a new project invitation","tags":["project-invitations"]}},"/project-invitations/process":{"post":{"operationId":"ProjectInvitationController_processInvitation","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProcessInvitationDto"}}}},"responses":{"200":{"description":"Invitation successfully processed"}},"security":[{"bearer":[]}],"summary":"Process (accept/reject) a project invitation","tags":["project-invitations"]}},"/project-invitations/project/{projectId}":{"get":{"operationId":"ProjectInvitationController_listProjectInvitations","parameters":[{"name":"projectId","required":true,"in":"path","schema":{"type":"string"}},{"name":"status","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns list of project invitations"}},"security":[{"bearer":[]}],"summary":"List all invitations for a project","tags":["project-invitations"]}},"/project-invitations/{id}/resend":{"post":{"operationId":"ProjectInvitationController_resendInvitation","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Invitation successfully resent"}},"security":[{"bearer":[]}],"summary":"Resend a project invitation","tags":["project-invitations"]}},"/project-invitations/{id}/cancel":{"put":{"operationId":"ProjectInvitationController_cancelInvitation","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Invitation successfully cancelled"}},"security":[{"bearer":[]}],"summary":"Cancel a project invitation","tags":["project-invitations"]}}},"info":{"title":"EV Charging Station Management System","description":"EV Charging Station Management System API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UpdateChargePointDto":{"type":"object","properties":{}},"CreateStationDto":{"type":"object","properties":{}},"UpdateStationDto":{"type":"object","properties":{}},"CreateLocationDto":{"type":"object","properties":{}},"UpdateLocationDto":{"type":"object","properties":{}},"CreateUserDto":{"type":"object","properties":{}},"UserResponse":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"ChangePasswordDto":{"type":"object","properties":{}},"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"sk@sk.com"},"firstName":{"type":"string","example":"Sonu"},"lastName":{"type":"string","example":"Kumar"},"password":{"type":"string","example":"123456!"},"phone":{"type":"string"},"role":{"type":"string","enum":["ADMIN","MANAGER","USER"],"default":"USER"}},"required":["email","firstName","lastName","password","role"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"skbhati199@gmail.com"},"password":{"type":"string","example":"12345678"}},"required":["email","password"]},"RefreshTokenDto":{"type":"object","properties":{}},"RequestPasswordResetDto":{"type":"object","properties":{}},"TestProjectInvitationDto":{"type":"object","properties":{"to":{"type":"string","example":"skbhati199@gmail.com"},"inviterName":{"type":"string","example":"sk"},"projectName":{"type":"string","example":"Project Demo"},"role":{"type":"string","example":"PROJECT_ADMIN"},"expiresAt":{"format":"date-time","type":"string","example":"2023-01-01"}},"required":["to","inviterName","projectName","role","expiresAt"]},"TestPasswordResetDto":{"type":"object","properties":{}},"CreateOrgDto":{"type":"object","properties":{"name":{"type":"string","example":"EV Technology Pvt. Ltd"},"email":{"type":"string","example":"sk@sk.com"},"phone":{"type":"string","example":"+919819658936"},"address":{"type":"string","example":"Devi Mandir, Khatriwara, Sikandrabad, UP 203205"}},"required":["name","email","phone","address"]},"UpdateOrgDto":{"type":"object","properties":{"name":{"type":"string","example":"EV Technology Pvt. Ltd"},"email":{"type":"string","example":"sk@sk.com"},"phone":{"type":"string","example":"+919819658936"},"address":{"type":"string","example":"Devi Mandir, Khatriwara, Sikandrabad, UP 203205"}},"required":["name","email","phone","address"]},"CreateProjectDto":{"type":"object","properties":{"name":{"type":"string","example":"Project Demo"},"description":{"type":"string","example":"This is a demo project"},"status":{"type":"string","example":"ACTIVE"},"organizationId":{"type":"string","example":"cm49ouxdp00003w73qder0ase"}},"required":["name","description","status","organizationId"]},"UpdateProjectDto":{"type":"object","properties":{"name":{"type":"string","example":"Project Demo"},"description":{"type":"string","example":"This is a demo project"},"status":{"type":"string","example":"ACTIVE"}},"required":["name","description","status"]},"AddProjectMemberDto":{"type":"object","properties":{"userId":{"type":"string","example":"User ID"},"role":{"type":"string","example":"PROJECT_ADMIN"}},"required":["userId","role"]},"CreateProjectInvitationDto":{"type":"object","properties":{"email":{"type":"string","example":"sk@sk.com"},"role":{"type":"string","example":"PROJECT_ADMIN"},"projectId":{"type":"string","example":"cm49ouxdp00003w73qder0ase"}},"required":["email","role","projectId"]},"ProcessInvitationDto":{"type":"object","properties":{"token":{"type":"string","example":"token"},"action":{"type":"string","example":"accept"}},"required":["token","action"]}}}}